Separated logic and integration
